fix(dashboard): infinite scrolling
In past behaviour, the raw scrolling event was fed and handled
incorrectly, firing multiple times and giving false positives,
triggering multiple API calls in the process.

A layer of indirection is added through a custom hook in order to
contain the event to just a single action that detects page's end.