ref(dashboard): split into multiple components
Some things are left to be thought about in the near future, specially
the state handling. Passing the setter obtained from useState hook is
probably a big no-no and it should be delegated either to a context or
lifting that up to the parent component. Both have arguments against;
context seems too heavy weight for this and lifting the state up would
mean handling the button differently.